% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_film_medata.R
\name{get_film_metadata}
\alias{get_film_metadata}
\title{Return the metadata of a film as a tibble.}
\usage{
get_film_metadata(url)
}
\arguments{
\item{url}{The film page URL on Allocine. It must match the right type of page / url : "https://www.allocine.fr/film/fichefilm_gen_cfilm=.*".}
}
\value{
The metadata of the film as a tibble.
}
\description{
Return the metadata of a film as a tibble.
}
\examples{
"https://www.allocine.fr/film/fichefilm_gen_cfilm=25802.html" \%>\%
  get_film_metadata()
}
