% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_press_ratings.R
\name{get_press_ratings}
\alias{get_press_ratings}
\title{Return the ratings of a film as a tibble.}
\usage{
get_press_ratings(url)
}
\arguments{
\item{url}{The ratings URL for a film on Allocine. It must match the right type of page / url : "^https://www.allocine.fr/film/fichefilm-.*/critiques/presse".}
}
\value{
The ratings of the film as a tibble.
}
\description{
Return the ratings of a film as a tibble.
}
\examples{
"https://www.allocine.fr/film/fichefilm-25802/critiques/presse/" \%>\%
  get_press_ratings()
}
