% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratings.R
\name{get_ratings}
\alias{get_ratings}
\title{Return ratings from all the films in a given year.}
\usage{
get_ratings(year, pages)
}
\arguments{
\item{year}{The year as a numeric value, e.g. 1999.}

\item{pages}{Only retrieve ratings from certain pages instead of all of them.}
}
\value{
The ratings and the metadata of the films as a tibble.
}
\description{
Return ratings from all the films in a given year.
}
\examples{
get_ratings(2017)
}
