% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Retrieve text of a node given a css attribute or an XPath request.}
\usage{
get_text(html, css, xpath, node = 1)
}
\arguments{
\item{html}{Either a document, a node set or a single node.}

\item{css, xpath}{Use either a css attribute or an XPath 1.0 request (not both).}

\item{node}{If several nodes match your selection, select the one you want. Defaults to 1 (first node).}
}
\value{
The text of the node.
}
\description{
Retrieve text of a node given a css attribute or an XPath request.
}
\examples{
html \%>\% get_text(".nationality")

film \%>\% get_text(xpath = "//*[contains(@class, 'actors')]//span[last()]")
}
