% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_filmlist.R
\name{process_filmlist}
\alias{process_filmlist}
\title{Return ratings from all the films in a given page.}
\usage{
process_filmlist(url)
}
\arguments{
\item{url}{The URL for a list of films on Allocine. It must match the right type of page / url : "^https://www.allocine.fr/films/decennie-\\d{4}/annee-\\d{4}/"".}
}
\value{
The ratings and the metadata of the films as a tibble.
}
\description{
Return ratings from all the films in a given page.
}
\examples{
"https://www.allocine.fr/films/decennie-1990/annee-1999/?page=1" \%>\%
  process_filmlist()
}
